# tests/Swagger/mollie.yaml
openapi: "3.0.0"
info:
  title: "Mollie Shopware Plugin API"
  description: "This is a Swagger Template for all uses cases of the Mollie plugin.<br /><br />Instructions:<ol><li>Configure the correct server domain, where your Shopware API ist available.</li><li>Then use the authorize request for Shopware to retrieve an Access-Token. Use the default credentials or just enter your custom credentials.</li><li>Just copy the Access-Token and use it in the global authorize section above the request list.</li></ol><br />Congratulations, you can now use the API requests."
servers:
  - url: /
  - url: https://localhost
  - url: http://localhost
  - url: '{custom}'
    variables:
      custom:
        default: https://localhost
...
paths:
...
  /api/mollie/ship/order?number={number}:
    get:
      tags:
        - "Shipping (Operational)"
      summary: "Full shipment"
      security:
        - AdminAPI: []
      parameters:
        - name: "number"
          in: "path"
          description: "Shopware order number"
          required: true
      responses:
        "200":
          description: "successful operation"
  /api/mollie/ship/item?order={order}&item={item}&quantity={quantity}:
    get:
      tags:
        - "Shipping (Operational)"
      summary: "Partial shipment"
      security:
        - AdminAPI: []
      parameters:
        - name: "order"
          in: "path"
          description: "Shopware order number"
          required: true
        - name: "item"
          in: "path"
          description: "product number"
        - name: "quantity"
          in: "path"
          description: "quantity"
      responses:
        "200":
          description: "successful operation"
...
components:
  securitySchemes:
    AdminAPI:
      type: apiKey
      in: header
      name: Authorization